More work, including fixing do not merge issue
