+++ src/database/googleSheetsDatabase.service.ts
@@ -11,68 +11,44 @@ import { getConfig } from '../config';
 import { Point } from '../point';
 import { Item } from '../item';
 import { Font } from '../font';
-import { Logger } from '@nestjs/common';
+import { Logger, Provider } from '@nestjs/common';
 import { DatabaseService } from './database.service';
 import { K } from 'handlebars';
 
-export const prodGoogleSheetsProvider = {
-  provide: 'DATABASE',
-  useFactory: async () => {
-    let config = await getConfig();
-    let creds = JSON.parse(
-      await readFile(__dirname + '/../../config/' + config['keyfile'], 'utf8'),
-    );
+// - testing_keyfile relative to config directory.
+export function createGoogleSheetsProvider(
+  testing_keyfile: string,
+  testing_spreadsheet_id: string,
+  readonly: boolean = false,
+): Provider {
+  return {
+    provide: 'DATABASE',
+    useFactory: async () => {
+      let creds = JSON.parse(
+        await readFile(__dirname + '/../../config/' + testing_keyfile, 'utf8'),
+      );
 
-    // Initialize auth - see https://theoephraim.github.io/node-google-spreadsheet/#/guides/authentication
-    const serviceAccountAuth = new JWT({
-      // env var values here are copied from service account credentials generated by google
-      // see "Authentication" section in docs for more info
-      email: creds.client_email,
-      key: creds.private_key,
-      scopes: ['https://www.googleapis.com/auth/spreadsheets'],
-    });
+      // Initialize auth - see https://theoephraim.github.io/node-google-spreadsheet/#/guides/authentication
+      const serviceAccountAuth = new JWT({
+        // env var values here are copied from service account credentials generated by google
+        // see "Authentication" section in docs for more info
+        email: creds.client_email,
+        key: creds.private_key,
+        scopes: [
+          readonly
+            ? 'https://www.googleapis.com/auth/spreadsheets.readonly'
+            : 'https://www.googleapis.com/auth/spreadsheets',
+        ],
+      });
 
-    let sheet_service = new GoogleSheetsService(
-      new GoogleSpreadsheet(config['spreadsheet-id'], serviceAccountAuth),
-    );
-    await sheet_service.init();
-    return sheet_service;
-  },
-};
-
-export const testGoogleSheetsProvider: {
-  provide: 'DATABASE';
-  useFactory: any;
-} = {
-  provide: 'DATABASE',
-  useFactory: async () => {
-    let config = await getConfig();
-    let creds = JSON.parse(
-      await readFile(
-        __dirname + '/../../config/' + config['testing-keyfile'],
-        'utf8',
-      ),
-    );
-
-    // Initialize auth - see https://theoephraim.github.io/node-google-spreadsheet/#/guides/authentication
-    const serviceAccountAuth = new JWT({
-      // env var values here are copied from service account credentials generated by google
-      // see "Authentication" section in docs for more info
-      email: creds.client_email,
-      key: creds.private_key,
-      scopes: ['https://www.googleapis.com/auth/spreadsheets'],
-    });
-
-    let sheet_service = new GoogleSheetsService(
-      new GoogleSpreadsheet(
-        config['testing-spreadsheet-id'],
-        serviceAccountAuth,
-      ),
-    );
-    await sheet_service.init();
-    return sheet_service;
-  },
-};
+      let sheet_service = new GoogleSheetsService(
+        new GoogleSpreadsheet(testing_spreadsheet_id, serviceAccountAuth),
+      );
+      await sheet_service.init();
+      return sheet_service;
+    },
+  };
+}
 
 class GoogleSheetsService extends DatabaseService {
   getPlayerSchema(): Promise<string[]> {
@@ -448,7 +424,11 @@ class GoogleSheetsService extends DatabaseService {
               (row) =>
                 row.get('gameId') == gameId && row.get('playerId') == playerId,
             )
-            .sort((a, b) => a.get('datetime') - b.get('datetime'));
+            .sort(
+              (a, b) =>
+                new Date(a.get('datetime')).getTime() -
+                new Date(b.get('datetime')).getTime(),
+            );
           if (sortedRows.length) {
             // We can use ! to avoid the undefined check here because of the
             // length check above.
